Updated the Python API, modified the Hibernate Connector to drop sessions quickly, however still buggy

+++ fda-table-service/rest-service/src/main/java/at/tuwien/endpoints/TableEndpoint.java
@@ -4,6 +4,7 @@ import at.tuwien.api.database.table.*;
 import at.tuwien.entities.database.table.Table;
 import at.tuwien.exception.*;
 import at.tuwien.mapper.TableMapper;
+import at.tuwien.service.DatabaseService;
 import at.tuwien.service.MessageQueueService;
 import at.tuwien.service.TableService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+import javax.ws.rs.NotAllowedException;
 import java.security.Principal;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,14 +28,16 @@ import java.util.stream.Collectors;
 @CrossOrigin(origins = "*")
 @RestController
 @RequestMapping("/api/container/{id}/database/{databaseId}/table")
-public class TableEndpoint {
+public class TableEndpoint extends AbstractEndpoint {
 
     private final TableService tableService;
     private final MessageQueueService amqpService;
     private final TableMapper tableMapper;
 
     @Autowired
-    public TableEndpoint(TableService tableService, MessageQueueService amqpService, TableMapper tableMapper) {
+    public TableEndpoint(TableService tableService, DatabaseService databaseService, MessageQueueService amqpService,
+                         TableMapper tableMapper) {
+        super(tableService, databaseService);
         this.tableService = tableService;
         this.amqpService = amqpService;
         this.tableMapper = tableMapper;
@@ -41,12 +45,15 @@ public class TableEndpoint {
 
     @GetMapping
     @Transactional(readOnly = true)
-    @PreAuthorize("hasRole('ROLE_RESEARCHER') and hasPermission(#databaseId, 'TABLE_VIEW')")
     @Operation(summary = "List all tables", security = @SecurityRequirement(name = "bearerAuth"))
     public ResponseEntity<List<TableBriefDto>> findAll(@NotNull @PathVariable("id") Long containerId,
                                                        @NotNull @PathVariable("databaseId") Long databaseId,
                                                        Principal principal)
             throws DatabaseNotFoundException {
+        if (!hasDatabasePermission(containerId, databaseId, "TABLES_VIEW", principal)) {
+            log.error("Missing table view permission");
+            throw new NotAllowedException("Missing table view permission");
+        }
         return ResponseEntity.ok(tableService.findAll(containerId, databaseId, principal)
                 .stream()
                 .map(tableMapper::tableToTableBriefDto)
@@ -55,7 +62,6 @@ public class TableEndpoint {
 
     @PostMapping
     @Transactional
-    @PreAuthorize("hasRole('ROLE_RESEARCHER') and hasPermission(#databaseId, 'TABLE_CREATE')")
     @Operation(summary = "Create a table", security = @SecurityRequirement(name = "bearerAuth"))
     public ResponseEntity<TableBriefDto> create(@NotNull @PathVariable("id") Long containerId,
                                                 @NotNull @PathVariable("databaseId") Long databaseId,
@@ -63,6 +69,10 @@ public class TableEndpoint {
                                                 Principal principal)
             throws ImageNotSupportedException, DatabaseNotFoundException, TableMalformedException, AmqpException,
             TableNameExistsException, ContainerNotFoundException, UserNotFoundException {
+        if (!hasDatabasePermission(containerId, databaseId, "TABLE_CREATE", principal)) {
+            log.error("Missing table create permission");
+            throw new NotAllowedException("Missing table create permission");
+        }
         final Table table = tableService.createTable(containerId, databaseId, createDto, principal);
         amqpService.create(table);
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -72,13 +82,16 @@ public class TableEndpoint {
 
     @GetMapping("/{tableId}")
     @Transactional(readOnly = true)
-    @PreAuthorize("hasRole('ROLE_RESEARCHER') and hasPermission(#databaseId, 'TABLE_VIEW')")
     @Operation(summary = "Get information about table", security = @SecurityRequirement(name = "bearerAuth"))
     public ResponseEntity<TableDto> findById(@NotNull @PathVariable("id") Long containerId,
                                              @NotNull @PathVariable("databaseId") Long databaseId,
                                              @NotNull @PathVariable("tableId") Long tableId,
                                              Principal principal)
             throws TableNotFoundException, DatabaseNotFoundException, ContainerNotFoundException {
+        if (!hasTablePermission(containerId, databaseId, tableId, "TABLE_INFO", principal)) {
+            log.error("Missing table view permission");
+            throw new NotAllowedException("Missing table view permission");
+        }
         final Table table = tableService.findById(containerId, databaseId, tableId, principal);
         log.debug(table);
         TableDto tableDto = tableMapper.tableToTableDto(table);
@@ -88,19 +101,20 @@ public class TableEndpoint {
 
     @PutMapping("/{tableId}")
     @Transactional
-    @PreAuthorize("hasRole('ROLE_RESEARCHER') and hasPermission(#databaseId, 'TABLE_UPDATE')")
     @Operation(summary = "Update a table", security = @SecurityRequirement(name = "bearerAuth"))
-    public ResponseEntity<TableBriefDto> update(@NotNull @PathVariable("id") Long id,
+    public ResponseEntity<TableBriefDto> update(@NotNull @PathVariable("id") Long containerId,
                                                 @NotNull @PathVariable("databaseId") Long databaseId,
                                                 @NotNull @PathVariable("tableId") Long tableId,
                                                 Principal principal) {
-        // TODO
+        if (!hasTablePermission(containerId, databaseId, tableId, "TABLE_UPDATE", principal)) {
+            log.error("Missing table update permission");
+            throw new NotAllowedException("Missing table update permission");
+        }
         return ResponseEntity.unprocessableEntity().body(new TableBriefDto());
     }
 
     @DeleteMapping("/{tableId}")
     @Transactional
-    @PreAuthorize("hasRole('ROLE_DEVELOPER') or hasRole('ROLE_DATA_STEWARD')")
     @Operation(summary = "Delete a table", security = @SecurityRequirement(name = "bearerAuth"))
     @ResponseStatus(HttpStatus.OK)
     public void delete(@NotNull @PathVariable("id") Long containerId,
@@ -108,7 +122,11 @@ public class TableEndpoint {
                        @NotNull @PathVariable("tableId") Long tableId,
                        Principal principal)
             throws TableNotFoundException, DatabaseNotFoundException, ImageNotSupportedException,
-            DataProcessingException, ContainerNotFoundException {
+            DataProcessingException, ContainerNotFoundException, TableMalformedException {
+        if (!hasTablePermission(containerId, databaseId, tableId, "TABLE_DELETE", principal)) {
+            log.error("Missing table delete permission");
+            throw new NotAllowedException("Missing table delete permission");
+        }
         tableService.deleteTable(containerId, databaseId, tableId, principal);
     }
